Handle locale strings with parameter indices
Attempt to fix #1.

Adding `%d*%$?` to the regexes in `CreateHybridPattern` (and updating
the other functions, too) should fix it for this case, but only because
the parameter indices are in order.

This 'fix' would break things horribly if parameter index specifiers
were used to refer to parameters in another order, and the captured
values were actually used. These format strings only have one string
and one number placeholder, so the `CreateStrNumPatternMatcher`
function hacks around that.